+++ lab_2/src/core/BitSequence/BitSequence.hpp
@@ -13,13 +13,13 @@ protected:
     DynamicArray<Bit<T_word>>* bits;
 
     BitSequence() : bits(new DynamicArray<Bit<T_word>>(0)) {}
-    BitSequence(Bit<T_word>* items, int count) 
+    BitSequence(Bit<T_word>* items, size_t count) 
         : bits(new DynamicArray<Bit<T_word>>(items, count)) {}
     BitSequence(const BitSequence& other) 
         : bits(new DynamicArray<Bit<T_word>>(*other.bits)) {}
     BitSequence(const LinkedList<Bit<T_word>>& list) 
         : bits(new DynamicArray<Bit<T_word>>(list.GetLength())) {
-        for (int i = 0; i < list.GetLength(); i++) {
+        for (size_t i = 0; i < list.GetLength(); i++) {
             bits->Set(i, list.Get(i));
         }
     }
@@ -33,27 +33,25 @@ protected:
 public:
     Bit<T_word> GetFirst() const override;
     Bit<T_word> GetLast() const override;
-    Bit<T_word> Get(int index) const override;
-    int GetLength() const override;
-    Sequence<Bit<T_word>>* GetSubsequence(int startIndex, int endIndex) const;
+    Bit<T_word> Get(size_t index) const override;
+    size_t GetLength() const override;
+    Sequence<Bit<T_word>>* GetSubsequence(size_t startIndex, size_t endIndex) const;
 
 
     Sequence<Bit<T_word>>* Append(Bit<T_word> item);
     Sequence<Bit<T_word>>* Prepend(Bit<T_word> item);
-    Sequence<Bit<T_word>>* InsertAt(Bit<T_word> item, int index);
+    Sequence<Bit<T_word>>* InsertAt(Bit<T_word> item, size_t index);
     Sequence<Bit<T_word>>* Concat(Sequence<Bit<T_word>>* otherSequence);
 
 
 
-    Bit<T_word>& operator[](int index);
-    const Bit<T_word>& operator[](int index) const;
+    Bit<T_word>& operator[](size_t index);
+    const Bit<T_word>& operator[](size_t index) const;
     Sequence<Bit<T_word>>* operator&(const BitSequence& other) const;
     Sequence<Bit<T_word>>* operator|(const BitSequence& other) const;
-    Sequence<Bit<T_word>>* operator^(const BitSequence& other) const;
+    Sequence<Bit<T_word>>* operator^(const BitSequence& other) const; // хорош
     Sequence<Bit<T_word>>* operator~() const;
 };
 
 
-#include "BitSequenceGetMethodes.tpp"
-#include "BitSequenceInsertMethode.tpp"
-#include "BitSequenceOperators.tpp"
+#include "BitSequence.tpp"
